Update recipe view with data passed to activity

+++ Android/app/src/main/java/com/example/cookbook/models/Ingredient.java
@@ -4,10 +4,10 @@ import java.io.Serializable;
 
 public class Ingredient implements Serializable {
     public String name;
-    public float quantity;
+    public int quantity;
     public String measure;
 
-    public Ingredient(String name, float quantity, String measure) {
+    public Ingredient(String name, int quantity, String measure) {
         this.name = name;
         this.quantity = quantity;
         this.measure = measure;
